// appservice/src/KuduModels.d.ts
export declare interface DeployResult {
    id?: string;
    status?: number;
    statusText?: string;
    authorEmail?: string;
    author?: string;
    deployer?: string;
    message?: string;
    progress?: string;
    receivedTime?: Date;
    startTime?: Date;
    endTime?: Date;
    lastSuccessEndTime?: Date;
    complete?: boolean;
    active?: boolean;
    isTemp?: boolean;
    isReadonly?: boolean;
    url?: string;
    logUrl?: string;
    siteName?: string;
}

/**
 * An interface representing LogEntry.
 */
export declare interface LogEntry {
    logTime?: Date;
    id?: string;
    message?: string;
    type?: number;
    detailsUrl?: string;
}